test-requirements.txt
odoo_test_helper
odoo13-addon-storage_backend @ git+https://github.com/OCA/storage@refs/pull/68/head#subdirectory=setup/storage_backend
odoo13-addon-storage_backend_sftp @ git+https://github.com/OCA/storage@refs/pull/68/head#subdirectory=setup/storage_backend_sftp
